:rocket: Refactor registration validation logic and introduce FormValidator class
